more work on testing manual set
All pass except ones that are added by Javascript.  Had to change to parse full doc, not readability-summery doc, in order to get all the blockquotes. Readability seems to aggresively prune them out with it's content scoring algorithm. This leaves us open to false positives where we think tweets embedded in footers and headers are part of the content, but we have to make tradeoffs here.